Change settings to serialize to JSON, all viewmodels to vmlocator
This is a little more complex, as I also crushed settings and interfaces into one (JsonSettingsRepository) and moved some functionality to other services (TwitchChannelService). After removing those, I had to refactor existing classes to be able to access the new viewmodels identically to their old functionality. Some dialog calling is still kind of janky as I have to pass Close() to the service and pass it through the loaded event.